Small clean up and refactor
