+++ rest-api/web/src/main/java/org/eclipse/kapua/app/api/web/JaxbContextResolver.java
@@ -84,10 +84,6 @@ import org.eclipse.kapua.service.authorization.role.RolePermissionQuery;
 import org.eclipse.kapua.service.authorization.role.RolePermissionXmlRegistry;
 import org.eclipse.kapua.service.authorization.role.RoleQuery;
 import org.eclipse.kapua.service.authorization.role.RoleXmlRegistry;
-import org.eclipse.kapua.service.datastore.ChannelInfoXmlRegistry;
-import org.eclipse.kapua.service.datastore.ClientInfoXmlRegistry;
-import org.eclipse.kapua.service.datastore.DatastoreMessageXmlRegistry;
-import org.eclipse.kapua.service.datastore.MetricInfoXmlRegistry;
 import org.eclipse.kapua.service.datastore.client.model.InsertResponse;
 import org.eclipse.kapua.service.datastore.model.ChannelInfo;
 import org.eclipse.kapua.service.datastore.model.ChannelInfoListResult;
@@ -102,6 +98,10 @@ import org.eclipse.kapua.service.datastore.model.query.ChannelInfoQuery;
 import org.eclipse.kapua.service.datastore.model.query.ClientInfoQuery;
 import org.eclipse.kapua.service.datastore.model.query.MessageQuery;
 import org.eclipse.kapua.service.datastore.model.query.MetricInfoQuery;
+import org.eclipse.kapua.service.datastore.model.xml.ChannelInfoXmlRegistry;
+import org.eclipse.kapua.service.datastore.model.xml.ClientInfoXmlRegistry;
+import org.eclipse.kapua.service.datastore.model.xml.DatastoreMessageXmlRegistry;
+import org.eclipse.kapua.service.datastore.model.xml.MetricInfoXmlRegistry;
 import org.eclipse.kapua.service.device.call.kura.model.bundle.KuraBundles;
 import org.eclipse.kapua.service.device.call.kura.model.configuration.KuraDeviceConfiguration;
 import org.eclipse.kapua.service.device.call.kura.model.deploy.KuraDeploymentPackage;
