+++ service/datastore/api/src/main/java/org/eclipse/kapua/service/datastore/DatastoreObjectFactory.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2017 Eurotech and/or its affiliates and others
+ * Copyright (c) 2011, 2018 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -16,7 +16,6 @@ import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.service.datastore.model.ChannelInfoListResult;
 import org.eclipse.kapua.service.datastore.model.ClientInfoListResult;
 import org.eclipse.kapua.service.datastore.model.MessageListResult;
-import org.eclipse.kapua.service.datastore.model.Metric;
 import org.eclipse.kapua.service.datastore.model.MetricInfoListResult;
 import org.eclipse.kapua.service.datastore.model.query.ChannelInfoQuery;
 import org.eclipse.kapua.service.datastore.model.query.ClientInfoQuery;
@@ -99,12 +98,4 @@ public interface DatastoreObjectFactory extends KapuaObjectFactory {
      */
     MetricInfoListResult newMetricInfoListResult();
 
-    /**
-     * Returns a new Metric instance
-     *
-     * @param name
-     * @param value
-     * @return
-     */
-    Metric<?> newMetric(String name, Object value);
 }
