+++ service/datastore/api/src/main/java/org/eclipse/kapua/service/datastore/model/ClientInfoListResult.java
@@ -11,7 +11,7 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.datastore.model;
 
-import org.eclipse.kapua.service.datastore.ClientInfoXmlRegistry;
+import org.eclipse.kapua.service.datastore.model.xml.ClientInfoXmlRegistry;
 
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
