+++ service/datastore/api/src/main/java/org/eclipse/kapua/service/datastore/model/query/MessageQuery.java
@@ -11,19 +11,18 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.datastore.model.query;
 
+import org.eclipse.kapua.service.datastore.model.DatastoreMessage;
+import org.eclipse.kapua.service.datastore.model.xml.DatastoreMessageXmlRegistry;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 
-import org.eclipse.kapua.service.datastore.DatastoreMessageXmlRegistry;
-import org.eclipse.kapua.service.datastore.model.DatastoreMessage;
-
 /**
  * Datastore message schema query definition
- * 
- * @since 1.0
  *
+ * @since 1.0
  */
 @XmlRootElement(name = "query")
 @XmlAccessorType(XmlAccessType.PROPERTY)
