+++ service/datastore/api/src/main/java/org/eclipse/kapua/service/datastore/model/xml/DatastoreMessageXmlRegistry.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2017 Eurotech and/or its affiliates and others
+ * Copyright (c) 2011, 2018 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,40 +9,41 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.datastore;
+package org.eclipse.kapua.service.datastore.model.xml;
 
 import org.eclipse.kapua.locator.KapuaLocator;
-import org.eclipse.kapua.service.datastore.model.ChannelInfoListResult;
-import org.eclipse.kapua.service.datastore.model.query.ChannelInfoQuery;
+import org.eclipse.kapua.service.datastore.DatastoreObjectFactory;
+import org.eclipse.kapua.service.datastore.model.MessageListResult;
+import org.eclipse.kapua.service.datastore.model.query.MessageQuery;
 
 import javax.xml.bind.annotation.XmlRegistry;
 
 /**
- * Channel information xml registry
+ * Datastore message xml registry
  *
  * @since 1.0
  */
 @XmlRegistry
-public class ChannelInfoXmlRegistry {
+public class DatastoreMessageXmlRegistry {
 
     private static final KapuaLocator LOCATOR = KapuaLocator.getInstance();
     private static final DatastoreObjectFactory DATASTORE_OBJECT_FACTORY = LOCATOR.getFactory(DatastoreObjectFactory.class);
 
     /**
-     * Creates a {@link ChannelInfoListResult} instance
+     * Creates a {@link MessageListResult} instance
      *
      * @return
      */
-    public ChannelInfoListResult newChannelInfoListResult() {
-        return DATASTORE_OBJECT_FACTORY.newChannelInfoListResult();
+    public MessageListResult newDatastoreMessageListResult() {
+        return DATASTORE_OBJECT_FACTORY.newDatastoreMessageListResult();
     }
 
     /**
-     * Creates a {@link ChannelInfoQuery} instance.
+     * Creates a {@link MessageQuery} instance.
      *
      * @return
      */
-    public ChannelInfoQuery newQuery() {
-        return DATASTORE_OBJECT_FACTORY.newChannelInfoQuery(null);
+    public MessageQuery newQuery() {
+        return DATASTORE_OBJECT_FACTORY.newDatastoreMessageQuery(null);
     }
 }
